ai_service/requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
opencv-python-headless
